distributed fix

+++ NeoML/Python/neoml/Dnn/DnnDistributed.py
@@ -21,7 +21,7 @@ class DnnDistributed(PythonWrapper.DnnDistributed):
     """Single process, multiple threads distributed training.
     
     :param dnn: The dnn to learn distributed.
-    :type dnn: neoml.Dnn
+    :type dnn: neoml.Dnn or str
     :param type: Learn on cpu or gpu.
     :type type: str, ("cpu", "gpu"), default="cpu"
     :param count: Count of models to use.
@@ -32,25 +32,52 @@ class DnnDistributed(PythonWrapper.DnnDistributed):
     :type path: str, default="distributed.arch"
     """
     def __init__(self, dnn, type='cpu', count=0, devs=None):
-        path = str(uuid.uuid4())
+        is_dnn = isinstance(dnn, neoml.Dnn)
+        path = str(uuid.uuid4()) if is_dnn else dnn
         if type == 'cpu':
             if count < 1:
                 raise ValueError('`count` must be a positive number.')
-            dnn.store(path)
+            if is_dnn:
+                dnn.store(path)
             super().__init__(path, count)
         elif type == 'cuda':
             if devs is None:
                 if count < 1:
                     raise ValueError('`devs` or `count` must be specified.')
                 devs = list(range(count))
-            dnn.store(path)
+            if is_dnn:
+                dnn.store(path)
             super().__init__(path, devs)
         else:
             raise ValueError('`type` must be one of: "cpu", "cuda".')
-        os.remove(path)
+        if is_dnn:
+            os.remove(path)
+
+    def run(self, set_data):
+        """Runs the network.
+
+        :param set_data: A callback that takes a math_engine and thread number
+            as an argument. It must return a dictionary of input blobs for the
+            dnn on a given thread, this dictionary must be the same as for the
+            learn method of a dnn.
+        :type set_data: callable
+        """
+        self._run(set_data)
+
+    def run_and_backward(self, set_data):
+        """Runs the network and performs a backward pass with the input data.
+
+        :param set_data: A callback that takes a math_engine and thread number
+            as an argument. It must return a dictionary of input blobs for the
+            dnn on a given thread, this dictionary must be the same as for the
+            learn method of a dnn.
+        :type set_data: callable
+        """
+        self._run_and_backward(set_data)
 
     def learn(self, set_data):
-        """Performs one iteration of learning.
+        """Runs the network, performs a backward pass 
+        and updates the trainable weights.
 
         :param set_data: A callback that takes a math_engine and thread number
             as an argument. It must return a dictionary of input blobs for the
@@ -59,6 +86,11 @@ class DnnDistributed(PythonWrapper.DnnDistributed):
         :type set_data: callable
         """
         self._learn(set_data)
+
+    def train(self):
+        """Updates the trainable weights of all models (after run_and_backward).
+        """
+        self._train()
 
     def last_losses(self, layer_name):
         """Gets values of the loss function on the last step for all models.
@@ -69,6 +101,11 @@ class DnnDistributed(PythonWrapper.DnnDistributed):
         """
         return self._last_losses(layer_name)
 
+    def get_output(self, layer_name):
+        """Returns last blobs of `layer_name` for all models.
+        """
+        return self._get_output(layer_name)
+
     def save(self, path):
         """Serializes the trained network.
 
@@ -76,3 +113,15 @@ class DnnDistributed(PythonWrapper.DnnDistributed):
         :type path: str
         """
         return self._save(path)
+
+    @solver.setter
+    def solver(self, new_solver):
+        """Sets the optimizer for the layer's trainable parameters.
+        """
+        self.set_solver(new_solver._internal)
+
+    @property
+    def get_model_count(self):
+        """Gets the number of models in disitrbuted traning.
+        """
+        return self._get_model_count()
